+++ shared/src/commonMain/kotlin/com/fintrack/shared/feature/budget/domain/model/Budget.kt
@@ -1,6 +1,7 @@
 package com.fintrack.shared.feature.budget.domain.model
 
 import com.fintrack.shared.feature.category.domain.model.Category
+import com.ionspin.kotlin.bignum.decimal.BigDecimal
 import kotlinx.datetime.LocalDate
 
 data class Budget(
@@ -8,7 +9,7 @@ data class Budget(
     val accountIds: List<String>,
     val name: String,
     val categories: List<Category>,
-    val limit: Double,
+    val limit: BigDecimal,
     val isExpense: Boolean,
     val startDate: LocalDate,
     val endDate: LocalDate
