+++ shared/src/commonMain/kotlin/com/fintrack/shared/feature/settings/domain/util/NotificationService.kt
@@ -1,6 +1,7 @@
 package com.fintrack.shared.feature.settings.domain.util
 
 import com.fintrack.shared.feature.transaction.domain.model.Transaction
+import com.ionspin.kotlin.bignum.decimal.BigDecimal
 import kotlinx.datetime.LocalTime
 import kotlinx.datetime.LocalDate
 
@@ -11,8 +12,8 @@ interface NotificationService {
     fun requestPermission(callback: (Boolean) -> Unit)
     fun showTransactionNotification(transaction: Transaction)
     fun showBudgetAlertNotification(budgetName: String, threshold: Int)
-    fun showBillReminderNotification(billName: String, amount: Double)
-    fun scheduleBillReminder(billName: String, amount: Double, dueDate: LocalDate, daysBefore: Int)
+    fun showBillReminderNotification(billName: String, amount: BigDecimal)
+    fun scheduleBillReminder(billName: String, amount: BigDecimal, dueDate: LocalDate, daysBefore: Int)
     fun showSummaryNotification(title: String, content: String)
     fun scheduleSummaryNotification(time: LocalTime)
     fun cancelSummaryNotification()
